move categories controller
